posts/requirements.txt
Werkzeug==1.0.1
Flask==1.1.2
Flask-Cors==3.0.9
redis==3.5.3
mysql-connector-python==8.0.26
